Add access roles for target accounts
There were two challenges here:

1) CDK 'stack' definition when deploying to many accounts

CDK wants stacks to be defined at build time, but this is
unwieldy when targeting many stacks. To get around this, a Stack
parameter, which RR traditionally populates, is used to override
the Stack tag value.

2) Shared config/secrets

The role template needs access to the ARN of the deploy tools
account. This is private (though not secret) information. It's
unclear what the 'right' approach is here to share this across
stacks. As a pragmatic solution, keeping the repository private
means we can just inline this data.