Use configured bucket from environment locally
This change is intended to improve the DX of this repository by allowing the lambda running locally to use the INFRA stack S3 bucket with the "DEV" prefix. It makes specifying a bucket name mandatory and expects this to be made available in a local `.env` file as follows:

```
DATA_BUCKET_NAME=deploy-infra-github-lens-githublensdatabucketUNIQUENAME
```

There is only an INFRA stack for this application and the value of `DATA_BUCKET_NAME` should be that in the currently created stack.

Note: When this is deployed and a stack is created next we'll need to update the .env file in S3 to reference the S3 bucket created in the INFRA stack.