feat: Allow collection of data for one GH team
This project heavily relies on the GitHub API. This API is rate limited.

In an attempt to improve the local DX, in this change we add a new configuration property `GITHUB_TEAM`.
When set, we reduce the amount of API calls made to hopefully yield a faster feedback loop when developing locally.