Add caching for stacks
We currently do a few things to speed things up and reduce AWS Cfn
API calls (which often hit rate limits).

* do not crawl if the aggregate JSON data has already been written
for today unless the --force-crawl flag is set
* only crawl once every 24 hours

This commit extends the above by adding caching for stack-level
data. Stack data is written to the store (S3 in prod) and then
future crawls read from that when it is not stale - defined as
when the stack has not been updated since the S3 file was written.

The vast majority of our Cloudformation stacks are not updated each
day so we should see quite high cache hit rates here.