Extract and test getRecordSetDomains logic
This change moves the logic to get the domains for a particular
stack into a new file, and breaks the functions required up for
easier understanding of what's going on.

We also add a simple test to cover some of the logic involved in
pulling this information from a cloudformation stack api response.