Add last-modified date to the data retrieved from S3
This change adds a `lastModified` field to the object
returned from S3 using the `RetrievedObject` interface.

The last modified date is determined from the S3 metadata
for a retrieved object.

The intention is to improve the usefulness of the data
returned from the API, (if the data in S3 hasn't been
refreshed it's useful to know that!)