Make API private by moving to ALB/EC2
Security Groups (network restrictions) are used to limit API access
to within the VPC and peered Guardian networks - specifically the
VPN.

API Gateway can be used to achieve this too but it is difficult to
use a custom domain when doing so. See e.g.

https://georgemao.medium.com/enabling-private-apis-with-custom-domain-names-aws-api-gateway-df1b62b0ba7c

which requires an ALB for this.

If, as it seems, an ALB is required anyway, then using API Gateway
seems to only add complexity here.

The shift to EC2 is not strictly speaking required but provides the
following:

* consistency with Cloudformation Lens (nee cdk-metadata)
* ability to use our CDK GuEc2App pattern
* a larger response payload (lambda behind ALB is limited to 1mb)

The cost impact of switching to EC2 here is $7/month for a single
t4g.nano instance (not reserved).

Longer-term it may make sense to move to a monorepo and share CDK
and ALB with Cloudformation Lens and the Services Catalogue itself.