List services for teams
This is it, the big one, the one you've all been waiting for...

P+E Teams (from Galaxies) are combined with Cloudformation stacks
and Github Repo information. The link points are:

* the 'gu:repo' tag on Cfn stacks
* the 'owners' info from Github Lens, which lists Github Teams
  that are admins for a repo
* the 'primaryGithubTeam' property on Galaxy teams

Note, this does not attempt any caching yet so each request does a
lot of work. We should implement that in a follow-up.