chore: Remove redundant esbuild configuration
The GitHub Lens and Services API both run on EC2.

Unlike the lambda runtime, EC2 requires the AWS SDK to be present in the artifact bundle.

In this change we remove the `--external` flag from `esbuild` for these services.

It's worth noting this would be a no-op as we're using AWS SDK v3, which is on the `@aws-sdk/*` namespace,
that is, the output from `esbuild` already includes the AWS SDK.