Introduce Cloudquery(!)
Cloudquery is a great way to collect AWS, and other, sources of
data and push it into a destination - in this case Postgres.

This commit represents an initial spike and there will be plenty
of follow up work to tidy things up and broaden its behaviour, e.g.
by targeting additional accounts.

See also:

https://www.cloudquery.io/

Co-authored-by: Julia <julia.branke@guardian.co.uk>
Co-authored-by: Natasha <natasha.thrale@guardian.co.uk>
Co-authored-by: Akash <akash.askoolum@guardian.co.uk>