fix(cloudquery): Automatically apply migration
In #160 we performed a plugin update, however we did not perform the necessary migration. This is resulting in the following messages being witnessed in the logs:

```log
failed to sync v1 source aws: failed to write for aws (v15.7.0) -> postgresql (v3.0.0): failed to CloseAndRecv client: rpc error: code = Internal desc = Context done: context canceled and failed to wait for plugin: table aws_accessanalyzer_analyzers not found in postgres. make sure to run migrate
```

We _could_ manually perform a migration, however whilst we're still rolling out to production, let's automatically migrate for ease.

Though undesired, a failing migration isn't too bad, as we should be able to clear the database, and re-crawl to fill it back up.

See:
  - https://www.cloudquery.io/docs/advanced-topics/managing-versions#managing-plugin-versions
  - https://www.cloudquery.io/docs/reference/destination-spec#migrate_mode