Run Cloudquery in all of our AWS accounts
This uses the OU support that Cloudquery provides to query all of
our AWS accounts without having to manually specify them. The docs
on how this works are here:

https://www.cloudquery.io/docs/plugins/sources/aws/multi-account#aws-organizations

Cloudquery discovers accounts by querying AWS for the OU members,
and then assumes a role in each target accounts to query it.

Note, the process isn't entirely automatic; the cloudquery-access
role must be present in target accounts, and this is only present
in accounts the aws-account-setup manages. See:

https://github.com/guardian/aws-account-setup/pull/58