fix(cloudquery): Obtain GitHub credentials from Secrets Manager
It is not possible to use SecureString SSM Parameters within CloudFormation,
so switch to using Secrets Manager.

We have one secret on path `/INFRA/deploy/cloudquery/github-credentials`.

To inject these as secrets into a container, we need to reference it by ARN.
We shouldn't commit this ARN into source control, so use `SecretsManager.fromSecretPartialArn` to obtain it.
Rather strangely, this adds an IAM policy with a resource of:

```
"Fn::Join": [
  "",
  [
    "arn:",
    {
      "Ref": "AWS::Partition",
    },
    ":secretsmanager:",
    {
      "Ref": "AWS::Region",
    },
    ":",
    {
      "Ref": "AWS::AccountId",
    },
    ":secret:/TEST/deploy/cloudquery/github-credentials-??????",
  ],
],
```

Not sure _how_ this works, but it does!