Introduce Galaxies as Cloudquery source
Galaxies is the Guardian's source of truth for people, teams, and
streams. The plugin used lives here:

https://github.com/guardian/cq-source-galaxies

Galaxies does not (atm) support machine access to the API. Reading
directly from the bucket is a 'pragmatic' solution here - it is easy
to implement but vulnerable to changes in the model. The team do not
expect these in the medium term though.

An initial rate of once per day is used as this data doesn't change
very often.