feat(cli): Simplify the `list-tasks` command output
Currently, the `list-tasks` command returns a list of all tasks, their ARN, and tags.

This is quite a lot of information, and more than we typically want.
Typically, we only want to know the task names.

This change adds a `debug` flag to the `list-tasks` command.
When `true`, all the detail is printed.
When `false`, only the names are printed. This is the default.

Usage:

```sh
npm run cli --workspace cloudquery list-tasks -- --stack deploy --stage CODE --app service-catalogue [--debug | --no-debug]
```