chore(repocop): Enable eslint to ensure consistent code style
When we first introduced Prisma in #280, we saw eslint exiting with OOM issues,
and so we disabled eslint from the repocop directory.

It looks like this is related to the size of the generated Prisma client library.
Reducing the size of the client library to exactly the tables we want (currently github_ tables)
allows us to use eslint without suffering OOM issues.

This change adds the `@@ignore` attribute to all the `aws_` tables which:

> Add @@ignore to a model that you want to exclude from Prisma Client (for example, a model that you do not want Prisma users to update).
> Ignored models are excluded from the generated Prisma Client.
> – https://www.prisma.io/docs/reference/api-reference/prisma-schema-reference#ignore-1

https://www.prisma.io/docs/concepts/components/introspection#introspecting-only-a-subset-of-your-database-schema offers an alternative approach,
however it would likely create a change to the final database schema,
or at least require a careful migration.

Using this annotation should not impact the final database schema.
That is, no tables will be dropped via this change.