feat(cloudquery): Collect S3 buckets in the own task
We're seeing some S3 buckets not appearing in the database.
The logs don't show anything obvious, however as S3 data collection is part of the
`RemainingAwsData` task, it is likely that the logs are not easy to parse.

This change moves S3 bucket collection into its own task.
Hopefully this makes it easier to identify issues.