feat(data-audit): Query the database (#504)
* refactor: Move database connection string logic to `common`

There are now two consumers of this logic:
  1. RepoCop
  2. Data Audit

* feat(data-audit): Connect to the database

Perform a small query to the database to check that we can.

* fix(data-audit): Allow lambda to connect to database

* docs: Add SQL used to manually create database user

* fix(data-audit): Add the prisma client to the lambda artifact

* refactor(data-audit): `SELECT COUNT(*)`

We're only interested in the number of records, rather than the contents, so `SELECT COUNT(*)`.

See https://www.prisma.io/docs/concepts/components/prisma-client/aggregation-grouping-summarizing#count

* feat(data-audit): Check the number of accounts we have data for

Check that the number of accounts in the database, is the same number of accounts in the AWS Organisation.

* fix: Larger pageSize

Co-authored-by: Natasha <67543397+NovemberTang@users.noreply.github.com>

---------

Co-authored-by: Natasha <67543397+NovemberTang@users.noreply.github.com>