refactor(github-actions-usage): Parse workflows using `@actions/workflow-parser`
Swap the community maintained Schema Store parsing of GitHub Workflow files
for the GitHub authored `@actions/workflow-parser` NPM module.

Interestingly, the two approaches yield slightly different results:
- https://github.com/guardian/service-catalogue/blob/3f25b6c553e1ed2b192bce11cc15e6f25afa7239/.github/workflows/ecs-publish.yml
  - Schema Store fails to parse
  - `@actions/workflow-parser` able to parse
- https://github.com/guardian/grid/blob/e37e3acaeea198beb82896a143cc9c66d200aadc/.github/workflows/ci.yml
  - Schema Store able to parse
  - `@actions/workflow-parser` fails to parse

Anecdotally, this version also appears to be more performant.

`@actions/workflow-parser` is an ESM only module, and it was proving tricky to get Jest to work with it.
For this reason, the tests use Node's native test runner.