Refactor DEV environment
This change aims to reduce the length of the `start` script.

We no longer install `libpq`, for the `pg_dump` command.
Instead, a `db-copy` container will run `pg_dump` against CODE,
and write files to disk.

These files are then mounted to a special location in the Postgres container,
which results in them being executed on launch.