fix: Compile each project individually
The `common` package has a compile error:

```
TS2742: The inferred type of createPullRequest cannot be named without a reference to `octokit-plugin-create-pull-request/node_modules/@octokit/types`. This is likely not portable. A type annotation is necessary.
```

It looks like compiling at the root isn't granular enough, as the error wasn't reported during builds.

This change resolves the compilation error, and updates to compiling each project individually.