fix: Add explicit return type to `createPullRequest`
There is a compilation error currently:

```
TS2742: The inferred type of createPullRequest cannot be named without a reference to `octokit-plugin-create-pull-request/node_modules/@octokit/types`. This is likely not portable. A type annotation is necessary.
```

The return type of `composeCreatePullRequest` is a bit tricky to infer.

We're (currently) only using the `html_url` property, so update `createPullRequest` to return that,
and therefore the return type of `createPullRequest` becomes `Promise<string | undefined>`.