fix(repocop): Prisma migration for repocop db user
Moves the manually applied SQL script to create the repocop user,
and apply permissions, to a Prisma migration, so that it gets automatically applied.

Whilst it looks like a no-op, the exexution of:

```sql
GRANT SELECT ON ALL TABLES IN SCHEMA public TO repocop;
```

Will grant `repocop` access to the `view_repo_ownership` view,
which was lost in #1066.

Co-authored-by: <ashcorr@users.noreply.github.com>