feat: Update the CloudQuery CLI to v6.15.0
This is the latest in the v6 major.

v6.0.0 defaulted `skip_dependent_tables` to `true`,
to avoid accidentally skipping a crucial table we're setting `skip_dependent_tables` to `false`.

We can remove this once we've switched to an "allow list" of tables.

See https://github.com/cloudquery/cloudquery/blob/main/cli/CHANGELOG.md#600-2024-07-18

Co-authored-by: Kelvin Chappell <kelvin-chappell@users.noreply.github.com>