feat: Alarm on `EBSByteBalance` metric anomalies
Provisions an alarm for the `EBSByteBalance` metric of the RDS instance.
This is in reaction to a recent incident where we saw the database being unusable from Grafana.
In triaging this incident, we saw the `EBSByteBalance` metric appearing to correlate with performance issues.
With this alarm, we hope to catch issues earlier.

See #1471.