feat: Create a view for the availability dashboard
The availability dashboard has quite a few variables referencing things like
stack, stage, app values and asg names.

There is a lot of repetition across these variables,
which makes them difficult to maintain.

This change introduces a view that can be used to source all the necessary information
for the availability dashboard.