feat(cloudquery): Collect `github_releases` for `guardian/cdk`
We'd like to understand more about how https://github.com/guardian/cdk is being used.
We already know which version of GuCDK a service is using.
Collecting `github_releases` for this repository will allow us to determine the latest version of GuCDK available,
and thus how up-to-date the estate is.

This change collects `github_releases` for individual repositories rather than the entire organisation
as it's yet not known how expensive this would be on the GitHub API rate limits and CloudQuery row limits.