refactor: Remove the `ignoredTables` list
Since switching to an allow list of tables (see #1665, #1672),
the list of tables in the repository match the list of tables collected with CloudQuery.

The `ignoredTables` list was used to filter the `availableTables` list before sending it to CloudQuery.
There is no overlap between these two lists, therefore it isn't needed.

The `awsTables` list represents what we collect.