Move SAML table to a separate task
The SAML identities table is a GitHub enterprise-level operation
instead of an organisation-level one like the existing tables. Because
GitHub apps have to either be one or the other, we need to create a
separate set of tasks that use a different GitHub app to authorise the
API calls.

This PR splits out the SAML table.